refactor: move credentials TS Provider with other types
