+++ clients/go/admin/client.go
@@ -5,12 +5,7 @@ import (
 	"crypto/tls"
 	"errors"
 	"fmt"
-	"io/ioutil"
-	"strings"
 	"sync"
-
-	"golang.org/x/oauth2"
-	"golang.org/x/oauth2/clientcredentials"
 
 	"github.com/flyteorg/flyteidl/clients/go/admin/mocks"
 	"github.com/flyteorg/flyteidl/clients/go/admin/pkce"
@@ -104,37 +99,6 @@ func getAuthenticationDialOption(ctx context.Context, cfg *Config, tokenSourcePr
 
 	oauthTokenSource := NewCustomHeaderTokenSource(tSource, cfg.UseInsecureConnection, clientMetadata.AuthorizationMetadataKey)
 	return grpc.WithPerRPCCredentials(oauthTokenSource), nil
-}
-
-type ClientCredentialsTokenSourceProvider struct {
-	ccConfig clientcredentials.Config
-}
-
-func NewClientCredentialsTokenSourceProvider(ctx context.Context, cfg *Config,
-	clientMetadata *service.PublicClientAuthConfigResponse, tokenURL string) (TokenSourceProvider, error) {
-
-	secretBytes, err := ioutil.ReadFile(cfg.ClientSecretLocation)
-	if err != nil {
-		logger.Errorf(ctx, "Error reading secret from location %s", cfg.ClientSecretLocation)
-		return nil, err
-	}
-
-	secret := strings.TrimSpace(string(secretBytes))
-	scopes := cfg.Scopes
-	if len(scopes) == 0 {
-		scopes = clientMetadata.Scopes
-	}
-
-	return ClientCredentialsTokenSourceProvider{
-		ccConfig: clientcredentials.Config{
-			ClientID:     cfg.ClientID,
-			ClientSecret: secret,
-			TokenURL:     tokenURL,
-			Scopes:       scopes}}, nil
-}
-
-func (p ClientCredentialsTokenSourceProvider) GetTokenSource(ctx context.Context) (oauth2.TokenSource, error) {
-	return p.ccConfig.TokenSource(ctx), nil
 }
 
 // InitializeAuthMetadataClient creates a new anonymously Auth Metadata Service client.
